consolidate text utils
