+++ src/tui/event-decorators/command.js
@@ -1,6 +1,5 @@
 import { formatTimestamp } from '../ui-utils/entry-utils.js';
-import { TextLayout } from '../ui-utils/text-layout.js';
-import { sanitizeText } from '../ui-utils/line-formatter.js';
+import { TextLayout, sanitizeText } from '../ui-utils/text-utils.js';
 
 import { TRACE_ICONS } from '../constants/index.js';
