+++ src/tui/ui-utils/entry-utils.js
@@ -1,5 +1,5 @@
 import { getDecorator } from '../event-decorators/index.js';
-import { TextLayout } from './text-layout.js';
+import { TextLayout } from './text-utils.js';
 import { TRACE_TYPES, LAYOUT } from '../constants/index.js';
 
 export const formatTimestamp = (timestamp) => {
