+++ src/tui/views/log-viewer/LogViewer.js
@@ -9,7 +9,7 @@ import { LineRenderer } from './LineRenderer.js';
 import { useStdoutDimensions } from '../../hooks/useStdoutDimensions.js';
 import { useSelection } from '../../hooks/useSelection.js';
 import { SHORTCUTS, buildFooter } from '../../constants/index.js';
-import { isEnterKey } from '../../ui-utils/input-utils.js';
+import { isEnterKey } from '../../ui-utils/text-utils.js';
 
 const Entry = ({ item, selected }) =>
   React.createElement(
