+++ src/tui/views/project-selector/ProjectSelector.js
@@ -3,7 +3,7 @@ import { Box, Text, useInput } from 'ink';
 import { Select } from '@inkjs/ui';
 import { AppContainer } from '../AppContainer.js';
 import { SHORTCUTS, buildFooter } from '../../constants/index.js';
-import { isBackKey } from '../../ui-utils/input-utils.js';
+import { isBackKey } from '../../ui-utils/text-utils.js';
 import fs from 'fs-extra';
 import path from 'path';
 import os from 'os';
