+++ src/tui/views/trace-types-filter/TraceTypesFilterView.js
@@ -5,7 +5,7 @@ import { useStdoutDimensions } from '../../hooks/useStdoutDimensions.js';
 import { AppContainer } from '../AppContainer.js';
 import { DEFAULT_FILTERS } from '../../ui-utils/entry-utils.js';
 import { SHORTCUTS, buildFooter } from '../../constants/index.js';
-import { isBackKey } from '../../ui-utils/input-utils.js';
+import { isBackKey } from '../../ui-utils/text-utils.js';
 
 /**
  * Full screen trace type filter view using ink-ui MultiSelect.
